Removed the "Tap to Refresh" functionality, fixed a couple bugs
I removed the "Tap to Refresh" functionality. Instead the Text View is removed from the view when not being used and shown again when needed. Also, fixed an issue where the lastMotionY was not being reset in resetHeader() which was causing the menu to be "jumpy" when swiping near the bottom.